# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Anaya's Board - Ludo Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  coins: number;
  level: number;
  isSuspended: boolean;
  isAdmin: boolean;
  /** @nullable */
  lastDailyReward?: string | null;
  createdAt: string;
}

export interface UserSync {
  clerkId: string;
  email: string;
  displayName: string;
  avatarUrl?: string;
}

export interface DailyRewardResult {
  coins: number;
  coinsEarned: number;
}

export interface UserStats {
  wins: number;
  losses: number;
  matches: number;
  level: number;
  coins: number;
  winRate: number;
}

export interface LeaderboardEntry {
  rank: number;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  wins: number;
  level: number;
  coins: number;
}

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface RoomPlayer {
  clerkId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  color: string;
  isReady: boolean;
}

export interface Room {
  id: number;
  code: string;
  hostId: string;
  status: RoomStatus;
  maxPlayers: number;
  players: RoomPlayer[];
  createdAt: string;
}

export interface RoomInput {
  /**
     * @minimum 2
     * @maximum 4
     */
  maxPlayers: number;
}

export interface AdminUsersPage {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface SuspendInput {
  suspended: boolean;
  reason?: string;
}

export interface CoinsInput {
  amount: number;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  totalGames: number;
  totalRooms: number;
  totalCoinsInCirculation: number;
}

export type AdminListUsersParams = {
page?: number;
limit?: number;
search?: string;
};