# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Anaya's Board - Ludo Game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomPlayer } from './roomPlayer';
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  code: string;
  hostId: string;
  status: RoomStatus;
  maxPlayers: number;
  players: RoomPlayer[];
  createdAt: string;
}