# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Anaya's Board - Ludo Game API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  clerkId: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  coins: number;
  level: number;
  isSuspended: boolean;
  isAdmin: boolean;
  /** @nullable */
  lastDailyReward?: string | null;
  createdAt: string;
}