# lib/api-zod/src/generated/types/userSync.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Anaya's Board - Ludo Game API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSync {
  clerkId: string;
  email: string;
  displayName: string;
  avatarUrl?: string;
}